Implemented get weather data by specified location
Implemented:
- get weather data by specified location
Changed:
- screens organize
- location activity logic
- other changes

+++ app/src/main/kotlin/ru/mts/avpopo85/weathery/data/model/base/openWeatherMap/currentWeather/IOWMCurrentWeatherMainResponse.kt
@@ -7,7 +7,7 @@ interface IOWMCurrentWeatherMainResponse {
     val temperature: Double
 
     /**Atmospheric pressure (on the sea level, if there is no sea_level or grnd_level data), hPa*/
-    val atmosphericPressureInhPa: Int
+    val atmosphericPressureInhPa: Double
 
     /**Humidity, %*/
     val humidity: Int
